+++ src/components/shared/store/storeItemCard/StoreItemCardHeader.tsx
@@ -9,11 +9,15 @@ import { CustomIcon } from '@/components/shared'
 interface StoreItemCardHeaderProps {
   item: StoreItem
   purchased?: boolean
+  placeholdersProducts: string[]
+  index: number
 }
 
 export function StoreItemCardHeader({
   item,
   purchased = false,
+  placeholdersProducts,
+  index,
 }: StoreItemCardHeaderProps) {
   return (
     <header className="relative max-h-[289px] items-center justify-center overflow-hidden rounded bg-gray-700/40 p-5">
@@ -29,9 +33,12 @@ export function StoreItemCardHeader({
         src={item.cover_image}
         alt={item.name}
         fill
-        sizes="100%"
+        sizes="70vw"
         className="object-scale-down"
         draggable={false}
+        quality={60}
+        placeholder="blur"
+        blurDataURL={placeholdersProducts[index]}
         priority
       />